+++ frontend/src/main.jsx
@@ -1,10 +1,11 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import App from './App.jsx'
+import PaginaInicial from './pages/pagina-inicial.jsx'
+
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <PaginaInicial />
   </StrictMode>,
 )
